Добавлина возмоность взаимодействия с вкладкой корзина и избранное

+++ src/pages/api/deletePost.ts
@@ -9,29 +9,25 @@ import PostModal from '../../../models/Post'
 
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  await connectDB();
-  try {
-    const { id } = req.query
-      PostModal.findOneAndDelete(
-        {
-          _id: id
-        }
-      )
-      .then( (doc) => {
-        if(!doc){
-          return res.status(404).json({
-            message: "Статья не найдена"
-          })
-        }
-  
-        res.json({
-          message: 'Succec'
+  await connectDB()
+  if (req.method === 'DELETE') {
+    try {
+      const { id } = req.query
+        PostModal.findOneAndDelete(
+          {
+            _id: id
+          }
+        )
+        .then( (doc) => {
+          if(!doc){
+            return res.status(404).json({ message: "Статья не найдена" })
+          }
+          res.json({ message: 'Succec' })
         })
-      })
-  } catch (error) {
-    console.log(error)
-    res.status(500).json({
-      message: 'Что-то пошло не так ыы'
-    });
+    } catch (error) {
+      res.status(500).json({ message: 'Не удалось удалить пост' })
+    }
+  }else{
+    return res.status(500).json({ message: "Запрос не является DELETE" })
   }
 }
